git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@52632 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/rating/ratinglib.php
@@ -156,10 +156,12 @@ class RatingLib extends TikiDb_Bridge
 			return false;
 		}
 
-        $this->query(
-            'DELETE FROM `tiki_user_votings` WHERE `user` = ? AND `id` = ?',
-            array($user, $token)
-        );
+		if (!empty($user)) {
+			$this->query(
+				'DELETE FROM `tiki_user_votings` WHERE `user` = ? AND `id` = ?',
+				array($user, $token)
+			);
+		}
 
 		$this->query(
 			'INSERT INTO `tiki_user_votings` ( `user`, `ip`, `id`, `optionId`, `time` ) VALUES( ?, ?, ?, ?, ? )',
